+++ craft_parts/utils/maven/common.py
@@ -393,9 +393,10 @@ class MavenXMLError(BaseException):
     """An error encountered while parsing XML for Maven projects."""
 
     message: str
+    details: str | None = None
 
     def __str__(self) -> str:
-        return self.message
+        return f"{self.message}\n{self.details}"
 
 
 def _find_element(
@@ -409,13 +410,16 @@ def _find_element(
     :param element: The haystack to search.
     :param path: The needle to find in the haystack.
     :param namespaces: A mapping of namespaces to use during the search.
-    :raises _MavenXMLError: if the needle can't be found.
+    :raises MavenXMLError: if the needle can't be found.
     :return: The discovered element.
     """
     if (needle := element.find(path, namespaces)) is not None:
         return needle
 
-    raise MavenXMLError(message=f"Could not parse {path}.")
+    raise MavenXMLError(
+        message=f"Could not find path {path!r} in element {element.tag!r}",
+        details=f"Could not find path {path!r} in the following XML element:\n{_format_xml_str(element)}",
+    )
 
 
 def _get_element_text(element: ET.Element) -> str:
@@ -431,4 +435,13 @@ def _get_element_text(element: ET.Element) -> str:
     if (text := element.text) is not None:
         return text
 
-    raise MavenXMLError(message=f"No text field found on {element.tag}.")
+    raise MavenXMLError(
+        message=f"No text field found on {element.tag!r}",
+        details=f"No text field found on {element.tag!r} in the following XML element:\n{_format_xml_str(element)}",
+    )
+
+
+def _format_xml_str(element: ET.Element) -> str:
+    """Get a nicely-formatted string for displaying an XML element."""
+    ET.indent(element)
+    return ET.tostring(element).decode(errors="replace")