+++ tests/unit/utils/test_maven_utils.py
@@ -242,7 +242,13 @@ def test_find_element() -> None:
     assert find_result is not None
     assert find_result.text == "Howdy!"
 
-    with pytest.raises(MavenXMLError, match="Could not parse"):
+    expected_error = dedent("""\
+        Could not find path 'nope' in element 'foo'
+        Could not find path 'nope' in the following XML element:
+        <foo>
+          <bar>Howdy!</bar>
+        </foo>""")
+    with pytest.raises(MavenXMLError, match=expected_error):
         _find_element(element, "nope", {})
 
 
@@ -253,7 +259,13 @@ def test_get_element_text() -> None:
 
     element = ET.fromstring("<foo><bar>Howdy!</bar></foo>")  # noqa: S314
 
-    with pytest.raises(MavenXMLError, match="No text field"):
+    expected_error = dedent("""\
+        No text field found on 'foo'
+        No text field found on 'foo' in the following XML element:
+        <foo>
+          <bar>Howdy!</bar>
+        </foo>""")
+    with pytest.raises(MavenXMLError, match=expected_error):
         _get_element_text(element)